repair actions, merge main, and run benchmarks
- Merged main into current branch with conflict resolution.
- Repaired Build paper PDF workflow to support jules-* branches.
- Created Internal Benchmark workflow to update README.md automatically.
- Updated run_internal_benchmark.py with CI-friendly JSON output.
- Cleaned up build artifacts, binary files, and __pycache__ directories.
- Replaced jules.md with updated repository state and development roadmap.
- Validated D3Q27 cascaded solver implementation and CLI info command.
