+++ .github/workflows/build-paper-pdf.yml
@@ -6,6 +6,7 @@ name: Build paper PDF
       - main
       - master
       - openfoam-benchmark-compare
+      - jules-*
     paths:
       - 'paper/**'
       - '.github/workflows/build-paper-pdf.yml'
@@ -53,7 +54,7 @@ jobs:
             exit 0
           fi
           git commit -m "docs: update paper pdf [skip ci]"
-          git push
+          git push origin ${{ github.ref_name }}
 
       - name: Upload PDF artifact
         uses: actions/upload-artifact@v4
