--- CLI/cascaded_lbm.py
@@ -1,432 +1,0 @@
-import torch
-from typing import Dict
-
-from lbm_utils import D3Q27Lattice, _compute_force_coefficients
-
-
-class CascadedLBM:
-    """Tensor-product raw-moment collision for D3Q27."""
-
-    MOMENT_KEYS = [(a, b, c) for a in range(3) for b in range(3) for c in range(3)]
-    MOMENT_NAMES = [f"{a}{b}{c}" for a, b, c in MOMENT_KEYS]
-
-    @staticmethod
-    def build_moment_basis(ex, ey, ez):
-        """Build the 27-row tensor-product raw-moment basis once per lattice."""
-        ex = ex.to(dtype=torch.float32)
-        ey = ey.to(dtype=torch.float32)
-        ez = ez.to(dtype=torch.float32)
-        basis_rows = []
-        for a, b, c in CascadedLBM.MOMENT_KEYS:
-            basis_rows.append((ex ** a) * (ey ** b) * (ez ** c))
-        return torch.stack(basis_rows, dim=0)
-
-    @staticmethod
-    def compute_raw_moments(f, ex, ey, ez, basis=None):
-        """Transform populations to tensor-product raw moments."""
-        if basis is None:
-            basis = CascadedLBM.build_moment_basis(ex, ey, ez)
-        basis = basis.to(device=f.device, dtype=f.dtype)
-        return torch.tensordot(basis, f, dims=([1], [0]))
-
-    @staticmethod
-    def equilibrium_raw_moments(rho, ux, uy, uz, cs2=1/3):
-        """Equilibrium tensor-product moments for D3Q27."""
-        K_eq = {}
-
-        def m1d(order, u):
-            if order == 0:
-                return torch.ones_like(rho)
-            if order == 1:
-                return u
-            if order == 2:
-                return u * u + cs2
-            raise ValueError(f"Unsupported moment order: {order}")
-
-        for a, b, c in CascadedLBM.MOMENT_KEYS:
-            key = f"{a}{b}{c}"
-            K_eq[key] = rho * m1d(a, ux) * m1d(b, uy) * m1d(c, uz)
-
-        return K_eq
-
-    @staticmethod
-    def cascaded_relax(K, K_eq, s_nu, s_e, s_h):
-        K_post = {}
-
-        # Conserved moments
-        for key in ("000", "100", "010", "001"):
-            K_post[key] = K[key]
-
-        # Second-order moments
-        K_post["200"] = K["200"] + s_e * (K_eq["200"] - K["200"])
-        K_post["020"] = K["020"] + s_e * (K_eq["020"] - K["020"])
-        K_post["002"] = K["002"] + s_e * (K_eq["002"] - K["002"])
-        K_post["110"] = K["110"] + s_nu * (K_eq["110"] - K["110"])
-        K_post["101"] = K["101"] + s_nu * (K_eq["101"] - K["101"])
-        K_post["011"] = K["011"] + s_nu * (K_eq["011"] - K["011"])
-
-        # Higher-order moments
-        for key in CascadedLBM.MOMENT_NAMES:
-            if key in K_post:
-                continue
-            K_eq_val = K_eq.get(key, torch.zeros_like(K[key]))
-            K_post[key] = K[key] + s_h * (K_eq_val - K[key])
-
-        return K_post
-
-    @staticmethod
-    def build_moment_matrix(ex, ey, ez):
-        return CascadedLBM.build_moment_basis(ex, ey, ez)
-
-    @staticmethod
-    def moments_to_populations(K, moment_matrix_inv):
-        """Exact inverse transform: tensor-product moments → populations."""
-        K_flat = torch.stack([K[name] for name in CascadedLBM.MOMENT_NAMES], dim=0)
-        shape = K_flat.shape[1:]
-        f_flat = moment_matrix_inv @ K_flat.reshape(27, -1)
-        return f_flat.reshape(27, *shape)
-
-
-class D3Q27CascadedSolver:
-    """D3Q27 LBM solver with cascaded raw-moment collision."""
-
-    def __init__(self, config, device: torch.device, phys_config):
-        self.config = config
-        self.device = device
-        self.resolution = config.resolution
-        self.phys_config = phys_config()
-
-        self._setup_physics_constants()
-
-        # D3Q27 lattice
-        self.ex, self.ey, self.ez = D3Q27Lattice.get_vectors()
-        self.ex = self.ex.to(device).nan_to_num(1e-12, posinf=1e18, neginf=-1e18)
-        self.ey = self.ey.to(device).nan_to_num(1e-12, posinf=1e18, neginf=-1e18)
-        self.ez = self.ez.to(device).nan_to_num(1e-12, posinf=1e18, neginf=-1e18)
-        self.w = D3Q27Lattice.get_weights().to(device).nan_to_num(1e-12, posinf=1e18, neginf=-1e18)
-        self.opposite = D3Q27Lattice.get_opposite().to(device).nan_to_num(1e-12, posinf=1e18, neginf=-1e18)
-        self.moment_basis = CascadedLBM.build_moment_basis(self.ex, self.ey, self.ez).to(device=device, dtype=torch.float32)
-        self.moment_matrix_inv = torch.inverse(self.moment_basis)
-        self._moment_names = CascadedLBM.MOMENT_NAMES
-        self._moment_basis_cached = self.moment_basis
-
-        # Populations (27 for D3Q27)
-        self.f = torch.zeros(27, self.resolution, self.resolution, self.resolution, device=device)+1e-12
-        self.f_temp = torch.zeros_like(self.f)+1e-12
-        self.f_pre_stream = torch.empty_like(self.f)
-
-        # Structure of Arrays (SoA) layout matching GPULBMSolver interface
-        self.velocity_x = torch.zeros(self.resolution, self.resolution, self.resolution, device=device)+1e-12
-        self.velocity_y = torch.zeros(self.resolution, self.resolution, self.resolution, device=device)+1e-12
-        self.velocity_z = torch.zeros(self.resolution, self.resolution, self.resolution, device=device)+1e-12
-        self.pressure = torch.zeros(self.resolution, self.resolution, self.resolution, device=device)+1e-12
-
-        # Turbulence and vorticity fields
-        self.nu_turb = torch.zeros(self.resolution, self.resolution, self.resolution, device=device)
-        self.vorticity = torch.zeros(3, self.resolution, self.resolution, self.resolution, device=device)
-        self.q_criterion = torch.zeros(self.resolution, self.resolution, self.resolution, device=device)
-
-        # Convergence tracking
-        self.velocity_prev = torch.zeros(self.resolution, self.resolution, self.resolution, device=device)+1e-12
-
-        # Force accounting from bounce-back / momentum exchange
-        self.force_x_accum = torch.tensor(0.0, device=device)
-        self.force_z_accum = torch.tensor(0.0, device=device)
-        self.force_samples = 0
-        self.force_x_last = torch.tensor(0.0, device=device)
-        self.force_z_last = torch.tensor(0.0, device=device)
-
-        # Cascaded relaxation parameters (these could be made configurable)
-        self.s_nu = 1.0 / 0.6    # Viscosity relaxation
-        self.s_e = 1.2           # Energy relaxation
-        self.s_h = 1.6           # Higher order relaxation
-
-        self._initialize_equilibrium()
-
-    def _setup_physics_constants(self):
-        """Compute STABLE physics constants"""
-        # Lattice units - everything is O(1)
-        self.cs2 = 1.0 / 3.0
-
-        # Use the lattice-consistent freestream speed for the Mach number.
-        # For D3Q27, c_s = 1/3, so u = Ma * c_s.
-        u_lattice = self.config.mach_number / 3.0
-        L_lattice = self.resolution
-        
-        # Force reasonable Reynolds number for this grid
-        # Keep tau safely above 0.6 to avoid the near-singular regime that
-        # produced unstable forces in the benchmark case.
-        Re_max_stable = L_lattice ** 1.5
-        Re_target = min(self.config.reynolds_number, Re_max_stable * 0.25)
-        
-        print(f"Reynolds number adjusted: {self.config.reynolds_number} → {Re_target:.0f}")
-        
-        # Compute viscosity in lattice units
-        self.nu = u_lattice * L_lattice / Re_target
-        
-        # Relaxation time (must be > 0.5 for stability)
-        tau = 3.0 * self.nu + 0.5
-        self.phys_config.s_nu = 1.0 / tau
-        
-        print(f"Lattice viscosity: {self.nu:.6f}")
-        print(f"Relaxation time: {tau:.4f}")
-        print(f"Relaxation parameter: {self.phys_config.s_nu:.4f}")
-        
-        # Validate stability
-        if tau < 0.6:
-            print("⚠️  WARNING: tau < 0.6, expect instability!")
-        if u_lattice > 0.15:
-            print("⚠️  WARNING: u > 0.15, expect instability!")
-
-
-    def _initialize_equilibrium(self):
-        """Initialize with D3Q27 equilibrium"""
-        rho = 1.0
-        u_lattice = self.config.mach_number / 3.0
-        ux = u_lattice  # Lattice-consistent freestream speed.
-        uy, uz = 0.0, 0.0
-
-        for i in range(27):
-            eu = self.ex[i] * ux + self.ey[i] * uy + self.ez[i] * uz
-            u_sq = ux*ux + uy*uy + uz*uz
-            feq = self.w[i] * rho * (1.0 + 3.0*eu + 4.5*eu**2 - 1.5*u_sq)
-            self.f[i] = feq
-
-    def _compute_vorticity(self, ux, uy, uz):
-        """Compute vorticity omega = curl(u) [web:44]"""
-        h = self.config.lbm_config.grid_spacing
-        # Compute all gradients in physical space.
-        grad_ux = torch.gradient(ux, spacing=(h, h, h), dim=(0, 1, 2))  # Returns (dux/dx, dux/dy, dux/dz)
-        grad_uy = torch.gradient(uy, spacing=(h, h, h), dim=(0, 1, 2))
-        grad_uz = torch.gradient(uz, spacing=(h, h, h), dim=(0, 1, 2))
-
-        # Extract individual components
-        dux_dx, dux_dy, dux_dz = grad_ux
-        duy_dx, duy_dy, duy_dz = grad_uy
-        duz_dx, duz_dy, duz_dz = grad_uz
-
-        # Vorticity: curl(u)
-        omega_x = duz_dy - duy_dz  # ∂w/∂y - ∂v/∂z
-        omega_y = dux_dz - duz_dx  # ∂u/∂z - ∂w/∂x
-        omega_z = duy_dx - dux_dy  # ∂v/∂x - ∂u/∂y
-
-        return omega_x.nan_to_num(1e-12, posinf=1e18, neginf=-1e18), omega_y.nan_to_num(1e-12, posinf=1e18, neginf=-1e18), omega_z.nan_to_num(1e-12, posinf=1e18, neginf=-1e18)
-
-    def _compute_q_criterion(self, ux, uy, uz):
-        """Compute Q-criterion for vortex identification"""
-        # Strain rate tensor magnitude
-        S11, S22, S33, S12, S13, S23 = self._compute_strain_rate_tensor(ux, uy, uz)
-        S_mag_sq = S11**2 + S22**2 + S33**2 + 2.0*(S12**2 + S13**2 + S23**2)
-
-        # Vorticity (rotation rate) magnitude
-        omega_x, omega_y, omega_z = self._compute_vorticity(ux, uy, uz)
-        omega_mag_sq = omega_x**2 + omega_y**2 + omega_z**2
-
-        # Q-criterion: Q > 0 indicates vortex regions
-        Q = 0.5 * (omega_mag_sq - S_mag_sq).nan_to_num(1e-12, posinf=1e18, neginf=-1e18)
-
-        return Q
-
-    def _compute_strain_rate_tensor(self, ux, uy, uz):
-        """Compute strain rate tensor S_ij = 0.5*(du_i/dx_j + du_j/dx_i)"""
-        h = self.config.lbm_config.grid_spacing
-        # Velocity gradients in physical space.
-        dux_dx, dux_dy, dux_dz = torch.gradient(ux, spacing=(h, h, h), dim=(0, 1, 2))
-        duy_dx, duy_dy, duy_dz = torch.gradient(uy, spacing=(h, h, h), dim=(0, 1, 2))
-        duz_dx, duz_dy, duz_dz = torch.gradient(uz, spacing=(h, h, h), dim=(0, 1, 2))
-
-        # Strain rate tensor (symmetric)
-        S11 = dux_dx.nan_to_num(1e-12, posinf=1e18, neginf=-1e18)
-        S22 = duy_dy.nan_to_num(1e-12, posinf=1e18, neginf=-1e18)
-        S33 = duz_dz.nan_to_num(1e-12, posinf=1e18, neginf=-1e18)
-        S12 = 0.5 * (dux_dy + duy_dx).nan_to_num(1e-12, posinf=1e18, neginf=-1e18)
-        S13 = 0.5 * (dux_dz + duz_dx).nan_to_num(1e-12, posinf=1e18, neginf=-1e18)
-        S23 = 0.5 * (duy_dz + duz_dy).nan_to_num(1e-12, posinf=1e18, neginf=-1e18)
-
-        return S11, S22, S33, S12, S13, S23
-
-    def collide_stream(self, geometry_mask: torch.Tensor, steps: int = 100):
-        """D3Q27 cascaded collision with streaming"""
-        h = self.config.lbm_config.grid_spacing
-        dt = self.config.lbm_config.time_step
-        # Reset force accounting for each run and only average over the
-        # quasi-steady tail of the run to suppress startup transients.
-        # A last-quarter window is a better tradeoff than whole-run averaging
-        # for the benchmark cube, while keeping the cost identical.
-        self.force_x_accum = torch.tensor(0.0, device=self.device)
-        self.force_z_accum = torch.tensor(0.0, device=self.device)
-        self.force_samples = 0
-        sample_window = max(10, steps // 4)
-        sample_start = max(0, steps - sample_window)
-
-        for step in range(steps):
-            # === 1. Compute macroscopic variables ===
-            rho = torch.sum(self.f, dim=0).nan_to_num(1e-12, posinf=1e18, neginf=-1e18)
-            ux = torch.sum(self.f * self.ex.view(-1, 1, 1, 1), dim=0).nan_to_num(1e-12, posinf=1e18, neginf=-1e18) / (rho + 1e-12)
-            uy = torch.sum(self.f * self.ey.view(-1, 1, 1, 1), dim=0).nan_to_num(1e-12, posinf=1e18, neginf=-1e18) / (rho + 1e-12)
-            uz = torch.sum(self.f * self.ez.view(-1, 1, 1, 1), dim=0).nan_to_num(1e-12, posinf=1e18, neginf=-1e18) / (rho + 1e-12)
-
-            # === 2. Store pre-stream populations for bounce-back ===
-            self.f_pre_stream.copy_(self.f)
-
-            # === 3. Cascaded collision using raw tensor-product moments ===
-            K = CascadedLBM.compute_raw_moments(self.f, self.ex, self.ey, self.ez, basis=self._moment_basis_cached)
-            K_eq = CascadedLBM.equilibrium_raw_moments(rho, ux, uy, uz)
-
-            # Update relaxation parameter
-            self.s_nu = 1.0 / (3.0 * self.nu + 0.5)
-
-            # Cascaded relaxation
-            K_post = CascadedLBM.cascaded_relax(K, K_eq, self.s_nu, self.s_e, self.s_h)
-
-            # Transform back to populations
-            self.f = CascadedLBM.moments_to_populations(K_post, self.moment_matrix_inv)
-
-            # === 4. Streaming ===
-            u_lattice = self.config.mach_number * 0.10
-            u_sq = u_lattice * u_lattice
-            for i in range(27):
-                dx = int(self.ex[i].item())
-                dy = int(self.ey[i].item())
-                dz = int(self.ez[i].item())
-                self.f_temp[i] = torch.roll(self.f[i], shifts=(dx, dy, dz), dims=(0, 1, 2))
-
-                # Prevent periodic wraparound at the outer domain boundary.
-                if dx > 0:
-                    self.f_temp[i][0, :, :] = self.f_pre_stream[i][0, :, :]
-                elif dx < 0:
-                    self.f_temp[i][-1, :, :] = self.f_pre_stream[i][-1, :, :]
-                if dy > 0:
-                    self.f_temp[i][:, 0, :] = self.f_pre_stream[i][:, 0, :]
-                elif dy < 0:
-                    self.f_temp[i][:, -1, :] = self.f_pre_stream[i][:, -1, :]
-                if dz > 0:
-                    self.f_temp[i][:, :, 0] = self.f_pre_stream[i][:, :, 0]
-                elif dz < 0:
-                    self.f_temp[i][:, :, -1] = self.f_pre_stream[i][:, :, -1]
-
-                # Open far-field treatment: refill wrapped populations with the
-                # same uniform freestream equilibrium used at initialization.
-                eu = self.ex[i] * u_lattice
-                feq_inf = self.w[i] * (1.0 + 3.0 * eu + 4.5 * eu * eu - 1.5 * u_sq)
-                if dx > 0:
-                    self.f_temp[i][:dx, :, :] = feq_inf
-                elif dx < 0:
-                    self.f_temp[i][dx:, :, :] = feq_inf
-                elif dy > 0:
-                    self.f_temp[i][:, :dy, :] = feq_inf
-                elif dy < 0:
-                    self.f_temp[i][:, dy:, :] = feq_inf
-                elif dz > 0:
-                    self.f_temp[i][:, :, :dz] = feq_inf
-                elif dz < 0:
-                    self.f_temp[i][:, :, dz:] = feq_inf
-
-            # === 5. Boundary conditions - bounce-back using pre-stream values ===
-            mask = geometry_mask > 0.5
-            step_force_x = torch.tensor(0.0, device=self.device)
-            step_force_z = torch.tensor(0.0, device=self.device)
-            for i in range(27):
-                opp_i = int(self.opposite[i].item())
-                self.f_temp[i] = torch.where(mask, self.f_pre_stream[opp_i], self.f_temp[i])
-
-                if i == 0 or i > opp_i:
-                    continue
-
-                dx = int(self.ex[i].item())
-                dy = int(self.ey[i].item())
-                dz = int(self.ez[i].item())
-                neighbor_is_solid = torch.roll(mask, shifts=(-dx, -dy, -dz), dims=(0, 1, 2))
-                boundary_link = (~mask) & neighbor_is_solid
-                if not torch.any(boundary_link):
-                    continue
-                step_force_x += torch.sum(2.0 * self.ex[i] * self.f_pre_stream[i][boundary_link])
-                step_force_z += torch.sum(2.0 * self.ez[i] * self.f_pre_stream[i][boundary_link])
-
-            self.force_x_last = step_force_x
-            self.force_z_last = step_force_z
-            if step >= sample_start:
-                self.force_x_accum += step_force_x
-                self.force_z_accum += step_force_z
-                self.force_samples += 1
-            self.f.copy_(self.f_temp)
-
-            # === 6. Update macroscopic fields for GUI interface ===
-            self.velocity_x = ux.nan_to_num(1e-12, posinf=1e18, neginf=-1e18)
-            self.velocity_y = uy.nan_to_num(1e-12, posinf=1e18, neginf=-1e18)
-            self.velocity_z = uz.nan_to_num(1e-12, posinf=1e18, neginf=-1e18)
-            self.pressure = (rho * self.cs2).nan_to_num(1e-12, posinf=1e18, neginf=-1e18)
-
-            # === 7. Compute vorticity and Q-criterion ===
-            if hasattr(self.phys_config, 'compute_q_criterion') and self.phys_config.compute_q_criterion:
-                omega_x, omega_y, omega_z = self._compute_vorticity(ux, uy, uz)
-                self.vorticity[0] = omega_x
-                self.vorticity[1] = omega_y
-                self.vorticity[2] = omega_z
-                self.q_criterion = self._compute_q_criterion(ux, uy, uz).nan_to_num(1e-12, posinf=1e18, neginf=-1e18)
-
-            # === 8. Diagnostic output ===
-            if step % 100 == 0:
-                if hasattr(self.vorticity, 'shape') and torch.sum(torch.isfinite(self.vorticity)) > 0:
-                    vorticity_mag = torch.sqrt(torch.sum(self.vorticity**2, dim=0))
-                    max_vort = torch.max(vorticity_mag.nan_to_num(0.0))
-                    print(f"Step {step}: max vorticity: {max_vort:.4f}, s_nu: {self.s_nu:.4f}")
-                else:
-                    print(f"Step {step}: s_nu: {self.s_nu:.4f}, vorticity not computed")
-
-            any_nan = torch.any(torch.isnan(self.f))
-            if any_nan:
-                print(f"WARNING: NaN detected at step {step}")
-                continue
-
-            if step % 500 == 0:
-                print(f"Step {step}/{steps} D3Q27 cascaded collision completed")
-
-    def compute_aerodynamic_coefficients(self, geometry_mask: torch.Tensor) -> Dict[str, float]:
-        """Compute forces using momentum-exchange method with enhanced diagnostics"""
-        h = self.config.lbm_config.grid_spacing
-
-        if self.force_samples > 0:
-            drag_force = self.force_x_accum / self.force_samples
-            lift_force = self.force_z_accum / self.force_samples
-            force_definition = 'bounce-back momentum exchange averaged over the last-quarter window'
-        else:
-            drag_force = self.force_x_last
-            lift_force = self.force_z_last
-            force_definition = 'bounce-back momentum exchange from last streaming step'
-
-        reference_area_voxelized = torch.sum(torch.any(geometry_mask > 0.5, dim=0).float()).item() * h**2
-        reference_area_voxelized = max(reference_area_voxelized, h**2)
-
-        coeffs = _compute_force_coefficients(
-            drag_force,
-            lift_force,
-            self.config.mach_number,
-            ref_area=reference_area_voxelized,
-            rho_ref=1.225,
-        )
-
-        # Basic diagnostics
-        rho = torch.sum(self.f, dim=0)
-        vorticity_mag = torch.sqrt(torch.sum(self.vorticity**2, dim=0)) if hasattr(self.vorticity, 'shape') else torch.zeros_like(rho)
-        v_inf = coeffs['freestream_speed']
-
-        return {
-            'force_x': drag_force.item(),
-            'force_z': lift_force.item(),
-            'drag_coefficient': coeffs['drag_coefficient'],
-            'lift_coefficient': coeffs['lift_coefficient'],
-            'force_definition': force_definition,
-            'reference_area': reference_area_voxelized,
-            'reference_area_voxelized': reference_area_voxelized,
-            'reference_length': h * self.resolution,
-            'freestream_speed': v_inf,
-            'density': coeffs['density'],
-            'pressure_sum': rho.sum().item(),
-            'max_turbulent_viscosity': self.nu,
-            'mean_smagorinsky_constant': 0.17,  # Default value
-            'max_vorticity': vorticity_mag.max().item() if hasattr(vorticity_mag, 'max') else 0.0,
-            'vortex_core_volume': 0.0,  # Not computed
-            'reynolds_number_turbulent': v_inf * h * self.resolution / self.nu
-        }
-
