--- CLI/mixed_precision_solver.py
@@ -1,100 +1,0 @@
-"""Mixed Precision LBM Solver - FP16 storage + FP32 compute
-Achieves 2-3x speedup on modern GPUs with minimal accuracy loss
-
-Usage:
-    from mixed_precision_solver import wrap_solver_mixed_precision
-    solver = GPULBMSolver(config, device, phys_config)
-    solver_fp16 = wrap_solver_mixed_precision(solver)
-"""
-import torch
-import warnings
-
-class MixedPrecisionWrapper:
-    def __init__(self, solver, enable_fp16=True, ddf_shift=True):
-        self.solver = solver
-        self.enable_fp16 = enable_fp16 and torch.cuda.is_available()
-        self.ddf_shift = ddf_shift
-        self.storage_dtype = torch.float16 if self.enable_fp16 else torch.float32
-        
-        if self.enable_fp16:
-            self._convert_to_fp16()
-            mem_saved = self._estimate_memory_savings()
-            print(f"Mixed Precision Enabled: {mem_saved:.1f}% memory reduction")
-    
-    def _estimate_memory_savings(self):
-        return 50.0 if self.enable_fp16 else 0.0
-    
-    def _convert_to_fp16(self):
-        if self.ddf_shift:
-            rho_ref = 1.0
-            ux = torch.zeros_like(self.solver.velocity_x)
-            uy = torch.zeros_like(self.solver.velocity_y)
-            uz = torch.zeros_like(self.solver.velocity_z)
-            
-            self.f_eq_ref = torch.zeros_like(self.solver.f, dtype=torch.float32)
-            for i in range(len(self.solver.ex)):
-                eu = self.solver.ex[i]*ux + self.solver.ey[i]*uy + self.solver.ez[i]*uz
-                u_sq = ux**2 + uy**2 + uz**2
-                self.f_eq_ref[i] = self.solver.w[i] * rho_ref * (1.0 + 3.0*eu + 4.5*eu**2 - 1.5*u_sq)
-            self.f_eq_ref = self.f_eq_ref.half()
-            self.solver.f = (self.solver.f - self.f_eq_ref.float()).half()
-        else:
-            self.solver.f = self.solver.f.half()
-    
-    def collide_stream(self, geometry_mask, steps=1):
-        for step in range(steps):
-            f_compute = self.solver.f.float()
-            if self.ddf_shift:
-                f_compute = f_compute + self.f_eq_ref.float()
-            
-            rho = torch.sum(f_compute, dim=0)
-            ux = torch.sum(f_compute * self.solver.ex.view(-1,1,1,1), dim=0) / (rho + 1e-12)
-            uy = torch.sum(f_compute * self.solver.ey.view(-1,1,1,1), dim=0) / (rho + 1e-12)
-            uz = torch.sum(f_compute * self.solver.ez.view(-1,1,1,1), dim=0) / (rho + 1e-12)
-            
-            for i in range(len(self.solver.ex)):
-                eu = self.solver.ex[i]*ux + self.solver.ey[i]*uy + self.solver.ez[i]*uz
-                u_sq = ux**2 + uy**2 + uz**2
-                feq = self.solver.w[i] * rho * (1.0 + 3.0*eu + 4.5*eu**2 - 1.5*u_sq)
-                omega = 1.0 / (3.0*self.solver.nu + 0.5)
-                f_compute[i] += omega * (feq - f_compute[i])
-            
-            f_pre = f_compute.clone()
-            for i in range(len(self.solver.ex)):
-                shifts = (int(self.solver.ex[i].item()), int(self.solver.ey[i].item()), int(self.solver.ez[i].item()))
-                f_compute[i] = torch.roll(f_compute[i], shifts=shifts, dims=(0,1,2))
-            
-            for i in range(len(self.solver.ex)):
-                opp_i = self.solver.opposite[i]
-                mask = geometry_mask > 0.5
-                f_compute[i] = torch.where(mask, f_pre[opp_i], f_compute[i])
-            
-            if self.ddf_shift:
-                self.solver.f = (f_compute - self.f_eq_ref.float()).half()
-            else:
-                self.solver.f = f_compute.half()
-            
-            self.solver.velocity_x = ux
-            self.solver.velocity_y = uy
-            self.solver.velocity_z = uz
-            self.solver.pressure = rho * self.solver.cs2
-    
-    def compute_macroscopic(self):
-        """Compute density and velocity"""
-        f_compute = self.solver.f.float()
-        if self.ddf_shift:
-            f_compute = f_compute + self.f_eq_ref.float()
-        
-        rho = torch.sum(f_compute, dim=0)
-        ux = torch.sum(f_compute * self.solver.ex.view(-1,1,1,1), dim=0) / (rho + 1e-12)
-        uy = torch.sum(f_compute * self.solver.ey.view(-1,1,1,1), dim=0) / (rho + 1e-12)
-        uz = torch.sum(f_compute * self.solver.ez.view(-1,1,1,1), dim=0) / (rho + 1e-12)
-        u = torch.stack([ux, uy, uz], dim=0)
-        return rho, u
-    
-    def __getattr__(self, name):
-        """Proxy all other attributes to the wrapped solver"""
-        return getattr(self.solver, name)
-
-def wrap_solver_mixed_precision(solver, enable_fp16=True):
-    return MixedPrecisionWrapper(solver, enable_fp16=enable_fp16)
