--- CLI/script.py
@@ -1,16 +1,0 @@
-
-# First, let me create a comprehensive requirements.txt to understand dependencies
-requirements = """torch>=2.0.0
-torchvision>=0.15.0
-numpy>=1.24.0
-scipy>=1.10.0
-scikit-image>=0.22.0
-click>=8.1.0
-pyyaml>=6.0
-tqdm>=4.65.0
-tensorboard>=2.13.0
-pydantic>=2.0.0
-"""
-
-print("Dependencies for the project:")
-print(requirements)
