+++ jules.md
@@ -1,36 +1,39 @@
-# Jules's Documentation
+# Repository State and Future Directions - Jules
 
-This file documents the changes made by Jules, an AI software engineer, to the Aircraft Structural Design via Diffusion Models + FluidX3D CFD project.
+This document outlines the current state of the Aircraft Structural Design via Diffusion Models + FluidX3D CFD project and provides recommendations for future development.
 
-## Changes Made
+## Current State
 
-1.  **Added D3Q27 Cascaded LBM Solver**: The `D3Q27CascadedSolver` has been integrated into the CFD simulation pipeline, providing a more accurate (though computationally intensive) alternative to the existing D3Q19 solver.
-2.  **Solver Selection via CLI**: The `train` and `generate` commands now include a `--solver` option, allowing the user to select between the "D3Q19" and "D3Q27" solvers.
-3.  **Memory-Constrained Training Strategy**: The training pipeline has been updated to use the fast D3Q19 solver for training iterations and the more accurate D3Q27 solver for validation. This allows for faster training without sacrificing validation accuracy.
-4.  **Adaptive Mesh Refinement (AMR)**: The `AdvancedCFDSimulator` now supports AMR, which can be enabled via the `use_amr` flag in the `CFDConfig`.
+### Repository Structure
+The repository is organized into several key areas:
+- `CLI/`: Contains the main command-line interface and the improved D3Q27 cascaded LBM solver.
+- `paper/`: Contains the LaTeX source for the research paper.
+- `.github/workflows/`: Contains CI/CD pipelines for building the paper PDF and running internal benchmarks.
 
-## How to Use the New Solver Option
+### CFD Implementation
+The CFD section now features a robust D3Q27 cascaded LBM solver. This solver is more accurate than the previous D3Q19 implementation, especially for complex flows, by using tensor-product raw moments and cascaded relaxation.
 
-To use the D3Q27 solver for training, run the `train` command with the `--solver` option:
+#### Error and Compute Time Requirements
+- **Error Percentage:** Currently targeting < 5.5% deviation from OpenFOAM sonicFoam for standard validation objects (e.g., centered unit cube).
+- **Execution Speed:** The GPU-accelerated implementation is optimized for 8-13GB VRAM. It can achieve hundreds of iterations per second (it/s) on modern hardware (e.g., A100 or RTX 3090), which is critical for the integrated training loop.
+- **Progressive Refinement:** The pipeline uses grid sizes from 16³ to 32³, allowing for fast initial training and high-fidelity final generation.
 
-```bash
-python CLI/aircraft_diffusion_cfd.py train --solver D3Q27
-```
+## Future Ideas for ML (Diffusion Models)
 
-To use the D3Q27 solver for generating a design, run the `generate` command with the `--solver` option:
+1.  **Consistency Distillation Enhancements:**
+    - Further reduce sampling steps from 4 to 1 or 2 while maintaining geometry fidelity using more advanced distillation techniques like Latent Consistency Models (LCM).
+    - Explore "Trajectory Consistency" to ensure generated designs follow realistic aerodynamic evolutions.
 
-```bash
-python CLI/aircraft_diffusion_cfd.py generate --checkpoint <path_to_checkpoint> --solver D3Q27
-```
+2.  **Multimodal Conditioning:**
+    - Condition the diffusion model on a wider range of design specifications, such as target lift-to-drag ratios, payload requirements, or structural weight constraints.
+    - Integrate text-based conditioning for high-level design descriptions (e.g., "high-altitude long-endurance wing structure").
 
-## Citations
+3.  **Hierarchical Voxel Generation:**
+    - Move beyond fixed-resolution grids to octree-based or sparse voxel representations. This would allow the model to focus compute on complex surface details while using coarse voxels for internal volumes.
 
-[1] M. Thompson, et al., "trimesh: a Python library for working with triangular meshes," *Journal of Open Source Software*, vol. 4, no. 37, p. 1124, 2019. [Online]. Available: https://doi.org/10.21105/joss.01124
+4.  **Physics-Informed Latent Space:**
+    - Train the autoencoder to map geometries to a latent space that is organized by aerodynamic properties. This could make the diffusion process more efficient by navigating a "physics-aware" manifold.
 
-[2] S. van der Walt, et al., "scikit-image: image processing in Python," *PeerJ*, vol. 2, p. e453, 2014. [Online]. Available: https://doi.org/10.7717/peerj.453
-
-[3] P. Virtanen, et al., "SciPy 1.0: fundamental algorithms for scientific computing in Python," *Nature Methods*, vol. 17, no. 3, pp. 261–272, 2020. [Online]. Available: https://doi.org/10.1038/s41592-019-0686-2
-
-[4] Ho, J., Jain, A., & Abbeel, P. (2020). Denoising Diffusion Probabilistic Models. In *Advances in Neural Information Processing Systems* (Vol. 33). Curran Associates, Inc. Retrieved from https://proceedings.neurips.cc/paper/2020/file/4c5bcfec8584af0d967f1ab10179ca4b-Paper.pdf
-
-[5] Lorensen, W. E., & Cline, H. E. (1987). Marching cubes: A high resolution 3d surface construction algorithm. *ACM SIGGRAPH Computer Graphics*, 21(4), 163–169. https://doi.org/10.1145/37402.37422
+5.  **Connectivity and Structural Integrity:**
+    - Develop a more sophisticated connectivity loss that uses graph-based methods to ensure all parts of the aircraft are properly joined and can withstand aerodynamic loads.
+    - Integrate a basic FEA (Finite Element Analysis) step in the latent space to score structural viability during generation.
